// common/model/track-ids.ts
export enum TrackIds {
	Melbourne = 0,
	PaulRicard = 1,
	Shanghai = 2,
	Bahrain = 3,
	Catalunya = 4,
	Monaco = 5,
	Montreal = 6,
	Silverstone = 7,
	Hockenheim = 8,
	Hungaroring = 9,
	Spa = 10,
	Monza = 11,
	Singapore = 12,
	Suzuka = 13,
	AbuDhabi = 14,
	Texas = 15,
	Brazil = 16,
	Austria = 17,
	Sochi = 18,
	Mexico = 19,
	Baku = 20,
	SakhirShort = 21,
	SilverstoneShort = 22,
	TexasShort = 23,
	SuzukaShort = 24,
	Hanoi = 25,
	Zandvoort = 26,
	Imola = 27,
	Portimao = 28,
	Jeddah = 29
}